requirements.txt
dash==3.0.4
dash_bootstrap_components==2.0.2
Flask==2.2.2
Flask_Login==0.6.3
folium==0.19.5
numpy==1.24.3
pandas==2.0.3
plotly==5.9.0
werkzeug==2.2.2
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.22.0